Update API Transactions

+++ database/migrations/2021_10_14_051301_create_menu_ms_table.php
@@ -22,7 +22,7 @@ class CreateMenuMsTable extends Migration
             $table->string('category')->nullable();
             $table->string('ingredients')->nullable();
             $table->integer('price')->nullable();
-            $table->numeric('ratingMenu')->nullable();
+            $table->float('ratingMenu', 8, 2)->nullable();
             $table->string('picturePath')->nullable();
 
             $table->string('is_active'); // active or not active for food or etc
